requirements.txt
coverage==4.4
Django==1.11.2
et-xmlfile==1.0.1
jdcal==1.2
pika==0.10.0
openpyxl==2.3.2
Pillow==3.1.0
psycopg2==2.6.2
reportlab==3.2.0
django-analytical==2.2.2
django-localflavor==1.3
django-statici18n==1.1.5
factory_boy==2.8.1
django-dotenv==1.4.1
pendulum==1.2.0
djangorestframework==3.3.2
gitpython==2.1.9
requests==2.18.3
django-bootstrap3==9.1.0
django-ordered-model==1.4.3
django-waffle==0.14.0
django-redis==4.8.0
beautifulsoup4==4.6.0
pypom==1.2.0
selenium==3.11.0
python-magic==0.4.15
django-ajax-selects==1.7.0
weasyprint==0.42.3

-r ./osis_common/requirements.txt
-r ./internship/requirements.txt
